DROID-3690 new push key logic
