+++ core-models/src/main/java/com/anytypeio/anytype/core_models/ObjectWrapper.kt
@@ -343,6 +343,8 @@ sealed class ObjectWrapper {
         val writersLimit: Double? by default
         val readersLimit: Double? by default
 
+        val spacePushNotificationEncryptionKey: String? by default
+
         val sharedSpaceLimit: Int
             get() {
                 val value = getValue<Double?>(Relations.SHARED_SPACES_LIMIT)
