+++ core-models/src/main/java/com/anytypeio/anytype/presentation/notifications/PushKey.kt
@@ -1,0 +1,10 @@
+package com.anytypeio.anytype.presentation.notifications
+
+data class PushKey(
+    val id: String,
+    val value: String
+) {
+    companion object {
+        val EMPTY = PushKey(value = "", id = "")
+    }
+}
