+++ domain/src/main/java/com/anytypeio/anytype/domain/notifications/PushKeyProvider.kt
@@ -1,0 +1,9 @@
+package com.anytypeio.anytype.domain.notifications
+
+import com.anytypeio.anytype.presentation.notifications.PushKey
+
+interface PushKeyProvider {
+    fun start()
+    fun stop()
+    fun getPushKey(): Map<String, PushKey>
+}
